Added to schema some things that Moral Kombat asked for
Added ways of storing pages and page information as requested by moral kombat, did not go with their method of storing due to the client requesting a different method specifically.  Arnesh and Jared helped with some of the logic and the ideas while working on this. Added them as co-authors

Co-Authored-By: ReganJ406 <54910371+ReganJ406@users.noreply.github.com>
Co-Authored-By: realarnesh <56060308+realarnesh@users.noreply.github.com>